Put Science Birds related code into a "package"
Put Science Birds related code into a Python "package".

NOTE: Putting a file named "__init__.py" under the package is required
in order to be able to import files (modules) from that directory
(package). That is, if we don't create a file named "__init__.py" under
a directory that contains some Python code, then we won't be able to
import anything from that directory (package).